fix #115 OFFCFIN(1) is stuck behind the console in cagepanel.

+++ Project1/gaEngine/gaNavMesh.h
@@ -66,7 +66,7 @@ namespace GameEngine {
 	public:
 		void addFloor(std::vector<std::vector<Point>>& polygons, float z);
 		void buildMesh(void);
-		float findPath(const glm::vec3& from, 
+		uint32_t findPath(const glm::vec3& from, 
 			const glm::vec3& to,
 			std::vector<glm::vec3>& path);						// build a path from/to
 	};
